+++ ui/src/pages/products/components/ColorLegend.tsx
@@ -1,5 +1,6 @@
 import { Box, Typography } from '@mui/material';
 import { Product7BoxBGColour } from './style/colors';
+import { NewReleases, NewReleasesOutlined } from '@mui/icons-material';
 
 interface LegendConfig {
   title: string;
@@ -19,7 +20,7 @@ const LEGEND_CONFIG: LegendConfig = {
       key: 'new',
       backgroundColor: Product7BoxBGColour.NEW,
       borderColor: Product7BoxBGColour.NEW,
-      description: 'New Concept',
+      description: 'New Proposed Concept',
     },
     {
       key: 'primitive',
@@ -54,8 +55,22 @@ const LEGEND_CONFIG: LegendConfig = {
   ],
 };
 
+const newLegendItems = [
+  {
+    key: 'newConceptTask',
+    icon: <NewReleases sx={{ fontSize: 'inherit' }} />,
+    description: 'New unpromoted concept in task',
+  },
+  {
+    key: 'newConceptProject',
+    icon: <NewReleasesOutlined sx={{ fontSize: 'inherit' }} />,
+    description: 'New unreleased concept in project',
+  },
+];
+
 export default function ColorLegend({
   showLegend = true,
+  includeNewConcept = true,
   config = LEGEND_CONFIG,
 }) {
   if (!showLegend) return null;
@@ -77,23 +92,44 @@ export default function ColorLegend({
         {config.title}
       </Typography>
 
-      {config.items.map(item => (
+      {newLegendItems.map(item => (
         <Box
           key={item.key}
           sx={{ display: 'flex', alignItems: 'center', gap: 0.5 }}
         >
-          <Box
+          <Typography
+            variant="caption"
+            component="span"
             sx={{
-              width: 12,
-              height: 12,
-              backgroundColor: item.backgroundColor,
-              border: `1px solid ${item.borderColor}`,
-              borderRadius: 1,
+              display: 'flex',
+              alignItems: 'center',
+              fontSize: 'inherit',
             }}
-          />
+          >
+            {item.icon}
+          </Typography>
           <Typography variant="caption">{item.description}</Typography>
         </Box>
       ))}
+      {config.items
+        .filter(item => item.key !== 'new' || includeNewConcept)
+        .map(item => (
+          <Box
+            key={item.key}
+            sx={{ display: 'flex', alignItems: 'center', gap: 0.5 }}
+          >
+            <Box
+              sx={{
+                width: 12,
+                height: 12,
+                backgroundColor: item.backgroundColor,
+                border: `1px solid ${item.borderColor}`,
+                borderRadius: 1,
+              }}
+            />
+            <Typography variant="caption">{item.description}</Typography>
+          </Box>
+        ))}
     </Box>
   );
 }